Updated EventService to provide regular and sticky events with a single request
The EventService now provides an object with two arrays, one for the
conventional events and one for the sticky events.
The StreamController has been updated to set the information on the
scope according to the changes in the EventService.
The Stream view has been updated to display the information accordingly
to how it is exposed by the controller.
Refs #11